Update navbar, remove logs, fix template loading error
