+++ web/src/layouts/MainLayout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react"
-import Navbar from "../components/Navbar"
+import Navbar from "../components/navbar/Navbar"
 
 
 interface IMainLayout {
