+++ web/src/pages/Home.tsx
@@ -8,9 +8,6 @@ import {ConnectWallet} from "../components/ConnectWallet"
 
 const Home = () => {
     const { session } = useStore(appStore)
-
-   
-
  
     return (
         <>
@@ -51,7 +48,7 @@ const Home = () => {
                         <Center h={'100%'}>
                             <Stack>
                                 <Text>Connect wallet to upload files</Text>
-                                <ConnectWallet  disConnectWebsocket={()=>{console.log('ws need global storage')}}/>
+                                <ConnectWallet />
                             </Stack>
                         </Center>
                     </Container>
